WIP fix new test/sample data fetching protocol

+++ packages/unzip_testdata.sh
@@ -25,6 +25,15 @@ mkdir -p ${ROOT_DATA_DIR}/swift
 mkdir -p ${ROOT_DATA_DIR}/chypp
 mv sf-test-data/swift_test_data.7z swift/
 mv sf-test-data/chypp_test_data.7z chypp/
+# First, let us check we have the sample data file installed. Otherwise no point.
+if [ ! -e ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/ ]; then
+    echo ERROR: ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/ not found
+    exit ${_ENOENT}; 
+fi
+if [ ! -e ${GITHUB_REPOS}/sf-test-data/swift_sample_data.rda ]; then
+    echo ERROR: 'swift_sample_data.rda' not found, expected ${GITHUB_REPOS}/sf-test-data/swift_sample_data.rda 
+    exit ${_ENOENT}; 
+fi
 mv sf-test-data/swift_sample_data.rda ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/
 
 cd ${ROOT_DATA_DIR}/swift/
@@ -34,22 +43,6 @@ chmod -R go+rx *
 cd ${ROOT_DATA_DIR}/chypp/
 7z x chypp_test_data.7z 
 chmod -R go+rx *
-
-# First, let us check we have the sample data file installed. Otherwise no point.
-if [ ! -e ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/ ]; then
-    echo ERROR: ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/ not found
-    exit ${_ENOENT}; 
-fi
-cd ${CSIRO_BITBUCKET}/swift/bindings/R/pkgs/swift/data/
-
-if [ ! -e swift_sample_data.rda ]; then
-    # curl -o swift_sample_data.rda https://cloudstor.aarnet.edu.au/plus/s/vfIbwcISy8jKQmg/download
-    if [ ! -e ${GITHUB_REPOS}/sf-test-data/swift_sample_data.rda ]; then
-        echo ERROR: 'swift_sample_data.rda' not found, and not found either in ${GITHUB_REPOS}/sf-test-data/swift_sample_data.rda 
-        exit ${_ENOENT}; 
-    fi
-    cp ${GITHUB_REPOS}/sf-test-data/swift_sample_data.rda ./
-fi
 
 if [ ! -e ${CHYPP_TEST_DIR} ]
 then
